requirements.txt
google-generativeai
pywin32
pyttsx3
Pillow
# Excel教学桌宠新增依赖
openai>=1.0.0
pandas>=2.0.0
openpyxl>=3.0.0
numpy>=1.24.0
ultralytics>=8.2.0
opencv-python>=4.8.0
